+++ docs/ui/components/Sidebar/SidebarHead.tsx
@@ -1,4 +1,5 @@
 import { DocsLogo, LinkBase } from '@expo/styleguide';
+import { PlanEnterpriseIcon } from '@expo/styleguide-icons/custom/PlanEnterpriseIcon';
 import { BookOpen02DuotoneIcon } from '@expo/styleguide-icons/duotone/BookOpen02DuotoneIcon';
 import { GraduationHat02DuotoneIcon } from '@expo/styleguide-icons/duotone/GraduationHat02DuotoneIcon';
 import { Home02DuotoneIcon } from '@expo/styleguide-icons/duotone/Home02DuotoneIcon';
@@ -44,6 +45,12 @@ export const SidebarHead = ({ sidebarActiveGroup }: SidebarHeadProps) => {
           title="Guides"
           Icon={BookOpen02DuotoneIcon}
           isActive={sidebarActiveGroup === 'general'}
+        />
+        <SidebarSingleEntry
+          href="/eas/"
+          title="EAS"
+          Icon={PlanEnterpriseIcon}
+          isActive={sidebarActiveGroup === 'eas'}
         />
         <SidebarSingleEntry
           href="/versions/latest/"