+++ packages/expo/ios/Fetch/NativeResponse.swift
@@ -36,9 +36,9 @@ internal final class NativeResponse: SharedObject, ExpoURLSessionTaskDelegate {
     self.dispatchQueue = dispatchQueue
   }
 
-  func startStreaming() {
+  func startStreaming() -> Data? {
     if isInvalidState(.responseReceived, .bodyCompleted) {
-      return
+      return nil
     }
     if state == .responseReceived {
       state = .bodyStreamingStarted
@@ -46,9 +46,9 @@ internal final class NativeResponse: SharedObject, ExpoURLSessionTaskDelegate {
       emit(event: "didReceiveResponseData", arguments: queuedData)
     } else if state == .bodyCompleted {
       let queuedData = self.sink.finalize()
-      emit(event: "didReceiveResponseData", arguments: queuedData)
-      emit(event: "didComplete")
+      return queuedData
     }
+    return nil
   }
 
   func cancelStreaming() {